refactored middleware and router; still broken
