--- STATE_OF_AFFAIRS.md
@@ -1,9 +1,0 @@
-I have no idea how to get an array inside an array to show up. 
-
-Thought I could just use a separate query, but that's not working. 
-
-One mongo query also doesn't work, because I can't get the stuff from other rows while mapping inside of another, so that's out. 
-
-No clue how to make this work. Problem lies inside of `recipe-model.js`.
-
-Don't copy this repo if your goal is to have a working product; its not working, its wonky, and you'll run into problems.
